+++ Scribe/Scribe-Api/src/utopia/scribe/api/database/access/logging/issue/variant/AccessIssueVariant.scala
@@ -18,7 +18,6 @@ object AccessIssueVariant extends AccessOneRoot[AccessIssueVariant[IssueVariant]
 	  * Access to individual issue variants in the DB, also including issue information
 	  */
 	lazy val withIssue = AccessIssueVariants.withIssues.head
-	
 	/**
 	  * Access to individual issue variants in the DB, also including issue occurrence information
 	  */
@@ -53,12 +52,10 @@ case class AccessIssueVariant[A](wrapped: TargetingOne[Option[A]])
 	  * A copy of this access which also targets issue
 	  */
 	lazy val joinIssue = join(ScribeTables.issue)
-	
 	/**
 	  * Access to the values of linked issue
 	  */
 	lazy val issue = AccessIssueValue(joinIssue)
-	
 	/**
 	  * Access to issue -based filtering functions
 	  */
@@ -68,12 +65,10 @@ case class AccessIssueVariant[A](wrapped: TargetingOne[Option[A]])
 	  * A copy of this access which also targets issue_occurrence
 	  */
 	lazy val joinOccurrence = join(ScribeTables.issueOccurrence)
-	
 	/**
 	  * Access to the values of linked issue occurrences
 	  */
 	lazy val occurrence = AccessIssueOccurrenceValue(joinOccurrence)
-	
 	/**
 	  * Access to issue occurrence -based filtering functions
 	  */