+++ Scribe/Scribe-Api/src/utopia/scribe/api/database/access/logging/issue/variant/FilterIssueVariants.scala
@@ -20,6 +20,11 @@ trait FilterIssueVariants[+Repr] extends FilterableView[Repr]
 	  */
 	def model = IssueVariantDbModel
 	
+	/**
+	  * @return Access to variants not caused by any error
+	  */
+	def notCausedByError = filter(model.errorId.isNull)
+	
 	
 	// OTHER	--------------------
 	
@@ -28,7 +33,6 @@ trait FilterIssueVariants[+Repr] extends FilterableView[Repr]
 	  * @return Copy of this access point that only includes issue variants with the specified version
 	  */
 	def affectingVersion(version: Version) = filter(model.version.column <=> version.toString)
-	
 	/**
 	  * @param versions Targeted versions
 	  * @return Copy of this access point that only includes issue variants where version is within the 
@@ -42,7 +46,6 @@ trait FilterIssueVariants[+Repr] extends FilterableView[Repr]
 	  * @return Copy of this access point that only includes issue variants with the specified issue id
 	  */
 	def ofIssue(issueId: Int) = filter(model.issueId.column <=> issueId)
-	
 	/**
 	  * @param issueIds Targeted issue ids
 	  * @return Copy of this access point that only includes issue variants where issue id is within the 
@@ -54,13 +57,21 @@ trait FilterIssueVariants[+Repr] extends FilterableView[Repr]
 	  * @param errorId error id to target
 	  * @return Copy of this access point that only includes issue variants with the specified error id
 	  */
-	def withError(errorId: Int) = filter(model.errorId.column <=> errorId)
-	
+	def causedBy(errorId: Int) = filter(model.errorId.column <=> errorId)
+	/**
+	  * @param errorId Id of the targeted error
+	  * @return Copy of this access limited to cases caused by the specified error.
+	  *         If 'errorId' was None, limits this access to cases not caused by any error.
+	  */
+	def causedBy(errorId: Option[Int]): Repr = errorId match {
+		case Some(errorId) => causedBy(errorId)
+		case None => notCausedByError
+	}
 	/**
 	  * @param errorIds Targeted error ids
 	  * @return Copy of this access point that only includes issue variants where error id is within the 
 	  * specified value set
 	  */
-	def withErrors(errorIds: IterableOnce[Int]) = filter(model.errorId.column.in(IntSet.from(errorIds)))
+	def causedByErrors(errorIds: IterableOnce[Int]) = filter(model.errorId.column.in(IntSet.from(errorIds)))
 }