+++ Scribe/Scribe-Core/src/utopia/scribe/core/model/combined/logging/ErrorRecordWithStackTrace.scala
@@ -2,8 +2,10 @@ package utopia.scribe.core.model.combined.logging
 
 import utopia.flow.collection.immutable.Empty
 import utopia.flow.view.template.Extender
+import utopia.scribe.core.model.factory.logging.ErrorRecordFactoryWrapper
 import utopia.scribe.core.model.partial.logging.ErrorRecordData
 import utopia.scribe.core.model.stored.logging.{ErrorRecord, StackTraceElementRecord}
+import utopia.vault.store.HasId
 
 /**
   * Combines error record data with stack trace information
@@ -13,7 +15,8 @@ import utopia.scribe.core.model.stored.logging.{ErrorRecord, StackTraceElementRe
 case class ErrorRecordWithStackTrace(error: ErrorRecord,
                                      stackTraceElements: Seq[StackTraceElementRecord] = Empty,
                                      cause: Option[ErrorRecordWithStackTrace] = None)
-	extends Extender[ErrorRecordData]
+	extends Extender[ErrorRecordData] with HasId[Int]
+		with ErrorRecordFactoryWrapper[ErrorRecord, ErrorRecordWithStackTrace]
 {
 	// COMPUTED --------------------------
 	
@@ -26,4 +29,7 @@ case class ErrorRecordWithStackTrace(error: ErrorRecord,
 	// IMPLEMENTED  ----------------------
 	
 	override def wrapped: ErrorRecordData = error.data
+	override protected def wrappedFactory: ErrorRecord = error
+	
+	override protected def wrap(factory: ErrorRecord): ErrorRecordWithStackTrace = copy(error = factory)
 }